Added basic core functionality to long_uri NEW and CREATE
The application roots to: long_uris_controller#new which is just a textfield
and a button for users to paste the link.

upon clicking the button it pushes a template against the model securing
an id that it will then use to encode in a base64 url-safe charset to be used
as suffix from then on to retrieve the original url by decoding back to decimal
as the id.

Drawbacks: two database operations for each record inserted. considering scalability
it might be a good idea to implement an apache foundations zookeeper system to provide
the bigint ids and keep track of everything. Or maybe swap postgre for redis.

config/initializer/constants.rb added to declare all coder-added related constants
app/helpers/long_uris_helper.rb contains key fucntions for encoding and decoding the
base64 ( ideas and most of the code from https://github.com/steventen/base62-rb )

additionally the helper contains a regex based function that will be used to extract
and inject the protocol out of the org_url string.