Implemented url validation top100 background job
app/jobs/top100_process_job.rb contains the required action for
populating the hits model but due to recent restrictions on heroku
(namely, cant make cron jobs and scheduler is a paid service now
https://devcenter.heroku.com/articles/scheduler ) it's not set up
to run at intervals so the feature is incomplete. But the job operates
correctly when there is enough data on the database. I just don't have a
way to schedule it

It's now set to call once, as soon as the 151th record is commited to the
db.