Added existent url feature
Instead of prompting an error when the user tries to enter an existing url to be shortened the algorithm will now transparently, and with a single select, redirect to the show action.

This feature will make it so if you already had a short url created but lost it, once you enter the original long url back in, you will get the existing short url that was initially created.