Reworked "Top Hits"
/top_hits now shows the list of the top 100 (if available) requested sites, ranked descending by the 'hits' counter

This feature uses an activejob job to query long_uri model and extract the top 100 requested uris.
It then proceeds to traverse the set and issue an HTTP GET request for each of the urls extracting only the title of each page.

This phase is just the basic functionality. The final polished feature will keep an in-memory top 100 and will update hits only
without re-hitting urls and only doing so for new entries. This way if an old url goes back on the top it will just retrieve it
from the backend instead of hitting the url for its <title> tag again.

bug fix:  long_uri create action will now include the host on the saved long url, 'www' was previously removed from it but it creates
issues when  agivensite.com and www.agivensite.com are present on the domain.