Added Nokogiri use to parse HTML title for top hits.
the line:

 title = Net::HTTP.get(URI.parse(url)).scan(/<title>(.*?)<\/title>/).first.flatten.join.slice(0,140)

was using regex to extract the first title found on any givem httm response body. this was raw html and
would not show special characters like  ' and " properly.

the replacement:

 title = Nokogiri::HTML(Net::HTTP.get(URI.parse(url))).css("title")[0].text

now invokes Nokogiri to parse the html and extract the first instance of the array for title which
is the actual tittle displayed on the browser were the uri invoked from it.

Added Procfile for heroku

Minor fixes on url generation of top_hits