Applied style to top_hits and improved hotfix for legacy org_urls having extraced its www. prefix by regex filter
Some of the urls tried to reach through NET::HTTP just return nil for title instead of previous open-uri errors so
a filter was needed for the edge cases of legacy so the job would preppend www. to the url before issuing the GET request

Note: still need to find a way to deal with session-dependant saved urls. Possible solution is a url flag the short url
gets replaced by a specific error template instead of trying to redirect to a failing url.